+++ back_end/saolei/userprofile/utils.py
@@ -1,0 +1,15 @@
+from captcha.models import CaptchaStore
+from django.utils import timezone
+
+# 验证验证码
+def judge_captcha(captchaStr, captchaHashkey):
+    if captchaStr and captchaHashkey:
+        get_captcha = CaptchaStore.objects.filter(
+            hashkey=captchaHashkey).first()
+        if get_captcha and get_captcha.response == captchaStr.lower():
+            # 图形验证码15分钟有效，get_captcha.expiration是过期时间
+            if (get_captcha.expiration - timezone.now()).seconds >= 0:
+                CaptchaStore.objects.filter(hashkey=captchaHashkey).delete()
+                return True
+    CaptchaStore.objects.filter(hashkey=captchaHashkey).delete()
+    return False
